requirements.txt
psycopg2-binary
sqlalchemy
pgvector
python-dotenv
google-genai>=1.0.0
pypdf
